Read the socket on the data event instead
There seems to be another bug in the way Node sockets work: if you
do what we did before (pause the socket and wait for the readable
event) your data is very much delayed. If you want to read data
in a timely manner, you must listen on the data event and do
buffering yourself.

Yay.